// frontend/src/api/model/AdminReceptionistDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Address } from "./Address";
import type { Instant } from "./Instant";
import type { LocalDate } from "./LocalDate";

export interface AdminReceptionistDto {
  publicId?: string;
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  /** @nullable */
  phoneNumber?: string | null;
  slug: string;
  icNumber: string;
  /**
   * @minLength 1
   * @maxLength 1
   */
  gender: string;
  dateOfBirth?: LocalDate;
  /** @nullable */
  avatarUrl?: string | null;
  address?: null | Address;
  createdAt?: Instant;
  deletedAt?: null | Instant;
}